+++ NEWS.md
@@ -1,6 +1,7 @@
 # bs4Dash 0.2.0.9000
 - update `bs4DashGallery()`. Add a "theme" argument
 - fix vignette titles
+- add a "fixed" argument to the `bs4DashNavbar()` (#7, thanks @theRcast)
 - update Readme
 
 # bs4Dash 0.1.0